adjusted trash can auto, fixed adjust bug

+++ src/Commands/Drive/DriveCommands/Move.cpp
@@ -1,6 +1,6 @@
 #include "Move.h"
 
-Move::Move(int a, float s, int time)
+Move::Move(int a, float s, double time)
 {
 	Requires(drivetrain);
 	speed = s;
